+++ kmath-functions/src/commonTest/kotlin/space/kscience/kmath/integration/GaussIntegralTest.kt
@@ -13,7 +13,7 @@ import kotlin.test.assertEquals
 class GaussIntegralTest {
     @Test
     fun gaussSin() {
-        val res = GaussIntegrator.integrate(0.0..2 * PI) { x ->
+        val res = GaussIntegrator.legendre(0.0..2 * PI) { x ->
             sin(x)
         }
         assertEquals(0.0, res.value!!, 1e-4)
@@ -21,7 +21,7 @@ class GaussIntegralTest {
 
     @Test
     fun gaussUniform() {
-        val res = GaussIntegrator.integrate(0.0..100.0,300) { x ->
+        val res = GaussIntegrator.legendre(0.0..100.0,300) { x ->
             if(x in 30.0..50.0){
                 1.0
             } else {